MILAB-6484: tidy impersonation code
- Restore the sharingSupported JSDoc: the impersonating getter had been
  inserted between the doc block and the getter it documents.
- Extract ensureProjectListRid into project_list.ts and use it in
  copyProjectToUser instead of inlining the resolve-or-create-list dance.
